Add ampRenderer
